workload/schemachange: refactor worker into op builder
Previously all operation generation logic was stateless and tied to the
schemaChangeWorker. Creating an independent builder struct will make it
easier to keep track of state when generating ops that relate to each other.

Release note: None
